requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
openai==1.3.0
pymupdf==1.23.3
faiss-cpu==1.7.4
numpy==1.25.2
pydantic==2.3.0
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1